+++ app/src/main/java/com/dscnsut/app/blogs/BlogDetailsFragment.kt
@@ -10,28 +10,8 @@ import android.widget.TextView
 import androidx.navigation.fragment.navArgs
 import com.dscnsut.app.R
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [fragment_blogDetails.newInstance] factory method to
- * create an instance of this fragment.
- */
 class fragment_blogDetails : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
 
     val safeArgs : fragment_blogDetailsArgs by navArgs()
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
@@ -48,24 +28,5 @@ class fragment_blogDetails : Fragment() {
 
 
         return view
-    }
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment fragment_blogDetails.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-                fragment_blogDetails().apply {
-                    arguments = Bundle().apply {
-                        putString(ARG_PARAM1, param1)
-                        putString(ARG_PARAM2, param2)
-                    }
-                }
     }
 }
